requirements.txt
bcrypt==4.2.1
cryptography==44.0.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Reuploaded==1.4.0
Flask-Limiter==3.10.1
mongoengine==0.29.1
PyJWT==2.10.1
pyotp==2.9.0
qrcode==8.0
pillow==11.1.0
Jinja2==3.1.5